Refactoring io subpackage
