+++ panel/io/model.py
@@ -1,0 +1,50 @@
+"""
+Utilities for manipulating bokeh models.
+"""
+from __future__ import absolute_import, division, unicode_literals
+
+from bokeh.models import Model
+from bokeh.protocol import Protocol
+
+from .state import state
+
+
+#---------------------------------------------------------------------
+# Public API
+#---------------------------------------------------------------------
+
+def diff(doc, binary=True, events=None):
+    """
+    Returns a json diff required to update an existing plot with
+    the latest plot data.
+    """
+    events = list(doc._held_events) if events is None else events
+    if not events or state._hold:
+        return None
+    msg = Protocol("1.0").create("PATCH-DOC", events, use_buffers=binary)
+    doc._held_events = [e for e in doc._held_events if e not in events]
+    return msg
+
+
+def remove_root(obj, replace=None):
+    """
+    Removes the document from any previously displayed bokeh object
+    """
+    for model in obj.select({'type': Model}):
+        prev_doc = model.document
+        model._document = None
+        if prev_doc:
+            prev_doc.remove_root(model)
+        if replace:
+            model._document = replace
+
+
+def add_to_doc(obj, doc, hold=False):
+    """
+    Adds a model to the supplied Document removing it from any existing Documents.
+    """
+    # Add new root
+    remove_root(obj)
+    doc.add_root(obj)
+    if doc._hold is None and hold:
+        doc.hold()
