+++ panel/tests/test_util.py
@@ -2,7 +2,7 @@ from collections import OrderedDict
 
 from bokeh.models import Div
 
-from panel.io import render_mimebundle
+from panel.io.notebook import render_mimebundle
 from panel.pane import PaneBase
 from panel.util import get_method_owner, abbreviated_repr
