+++ panel/viewable.py
@@ -7,8 +7,6 @@ from __future__ import absolute_import, division, unicode_literals
 
 import io
 import re
-import signal
-import uuid
 
 from functools import partial
 from six import string_types
@@ -21,13 +19,13 @@ from bokeh.io import curdoc as _curdoc, export_png as _export_png
 from bokeh.embed import file_html as _file_html
 from bokeh.resources import CDN as _CDN
 from bokeh.models import CustomJS
-from bokeh.server.server import Server
 from bokeh.util.string import decode_utf8
-from pyviz_comms import JS_CALLBACK, JupyterCommManager, Comm as _Comm
+from pyviz_comms import JupyterCommManager, Comm as _Comm
 
-from .io import (
-    ABORT_JS, add_to_doc, push, render_mimebundle, state, embed_state,
-    render_model, _origin_url, show_server, config)
+from .config import config
+from .io.notebook import (get_comm_customjs, push, render_mimebundle,
+                          render_model, show_server)
+from .io import add_to_doc, embed_state, get_server, state
 from .util import param_reprs
 
 
@@ -334,15 +332,7 @@ class Viewable(Layoutable):
         port: int (optional, default=0)
            Allows specifying a specific port
         """
-        if callable(notebook_url):
-            origin = notebook_url(None)
-        else:
-            origin = _origin_url(notebook_url)
-        server_id = uuid.uuid4().hex
-        server = self.get_server(port, origin, start=True, show=False,
-                                 server_id=server_id)
-        show_server(server, notebook_url, server_id)
-        return server
+        return show_server(self, notebook_url, port)
 
     def embed(self, max_states=1000, max_opts=3, json=False,
               save_path='./', load_path=None):
@@ -374,74 +364,10 @@ class Viewable(Layoutable):
                         json, save_path, load_path)
         publish_display_data(*render_model(model))
 
-    def get_server(self, port=0, websocket_origin=None, loop=None,
+    def _get_server(self, port=0, websocket_origin=None, loop=None,
                    show=False, start=False, **kwargs):
-        """
-        Returns a Server instance with this panel attached as the root
-        app.
-
-        Arguments
-        ---------
-        port: int (optional, default=0)
-           Allows specifying a specific port
-        websocket_origin: str or list(str) (optional)
-           A list of hosts that can connect to the websocket.
-
-           This is typically required when embedding a server app in
-           an external web site.
-
-           If None, "localhost" is used.
-        loop : tornado.ioloop.IOLoop (optional, default=IOLoop.current())
-           The tornado IOLoop to run the Server on
-        show : boolean (optional, default=False)
-           Whether to open the server in a new browser tab on start
-        start : boolean(optional, default=False)
-           Whether to start the Server
-        kwargs: dict
-           Additional keyword arguments to pass to Server instance
-
-        Returns
-        -------
-        server : bokeh.server.server.Server
-           Bokeh Server instance running this panel
-        """
-        from tornado.ioloop import IOLoop
-        opts = dict(kwargs)
-        if loop:
-            loop.make_current()
-            opts['io_loop'] = loop
-        else:
-            opts['io_loop'] = IOLoop.current()
-
-        if websocket_origin:
-            if not isinstance(websocket_origin, list):
-                websocket_origin = [websocket_origin]
-            opts['allow_websocket_origin'] = websocket_origin
-
-        server_id = kwargs.pop('server_id', None)
-        server = Server({'/': partial(self._modify_doc, server_id)}, port=port, **opts)
-        if server_id:
-            state._servers[server_id] = (server, self, [])
-
-        if show:
-            def show_callback():
-                server.show('/')
-            server.io_loop.add_callback(show_callback)
-
-        def sig_exit(*args, **kwargs):
-            server.io_loop.add_callback_from_signal(do_stop)
-
-        def do_stop(*args, **kwargs):
-            server.io_loop.stop()
-        signal.signal(signal.SIGINT, sig_exit)
-
-        if start:
-            server.start()
-            try:
-                server.io_loop.start()
-            except RuntimeError:
-                pass
-        return server
+        return get_server(self, port, websocket_origin, loop, show,
+                          start, **kwargs)
 
     def save(self, filename, title=None, resources=None, template=None,
              template_variables={}, embed=False, max_states=1000,
@@ -573,11 +499,11 @@ class Viewable(Layoutable):
             from .util import StoppableThread
             loop = IOLoop()
             server = StoppableThread(
-                target=self.get_server, io_loop=loop,
+                target=self._get_server, io_loop=loop,
                 args=(port, websocket_origin, loop, True, True))
             server.start()
         else:
-            server = self.get_server(port, websocket_origin, show=True, start=True)
+            server = self._get_server(port, websocket_origin, show=True, start=True)
 
         return server
 
@@ -721,19 +647,9 @@ class Reactive(Viewable):
         Returns a CustomJS callback that can be attached to send the
         model state across the notebook comms.
         """
-        # Abort callback if value matches last received event
-        abort = ABORT_JS.format(plot_id=plot_id, change=change)
-        data_template = "data = {{{change}: cb_obj['{change}'], 'id': cb_obj.id}};"
-        fetch_data = data_template.format(change=change)
-        self_callback = JS_CALLBACK.format(comm_id=client_comm.id,
-                                           timeout=self._timeout,
-                                           debounce=self._debounce,
-                                           plot_id=plot_id)
-        js_callback = CustomJS(code='\n'.join([abort,
-                                               fetch_data,
-                                               self_callback]))
-        return js_callback
-
+        return get_comm_customjs(change, client_comm, plot_id,
+                                 self._timeout, self._debounce)
+        
     #----------------------------------------------------------------
     # Model API
     #----------------------------------------------------------------
